Store EntityMetaData by EntityId
Currently unloading then reloading an Entity is making the TempMetaData persist across these transitions.

This makes hard to use TempMeta to know if something needs to be initialized or not (listeners) and is carrying
over old listeners to the old entity that no longer exists.

Using the EntityId + world name is guaranteed unique per server session and therefor will be unique to the living entity's object.